+++ platform-engineer-2/images.json
@@ -3,6 +3,7 @@
     "https://www.w3schools.com/w3css/img_lights.jpg",
     "https://www.w3schools.com/w3css/img_forest.jpg",
     "https://www.w3schools.com/w3css/img_mountains.jpg",
-    "https://www.w3schools.com/w3css/img_snowtops.jpg"
+    "https://www.w3schools.com/w3css/img_snowtops.jpg",
+    "https://images.pexels.com/photos/326807/pexels-photo-326807.jpeg?cs=srgb&dl=pexels-pixabay-326807.jpg&fm=jpg"
   ]
 }
