+++ platform-engineer-2/package.json
@@ -4,7 +4,7 @@
   "description": "Zarv Challenge",
   "main": "dist/index.js",
   "scripts": {
-    "test": "npm run test",
+    "test": "jest --verbose --coverage --runInBand --forceExit",
     "build": "./node_modules/typescript/bin/tsc",
     "start": "npm run build && node ./dist/index.js"
   },
@@ -22,8 +22,10 @@
   },
   "homepage": "https://github.com/domenester/zarv-challenge#readme",
   "devDependencies": {
+    "@types/jest": "^29.5.5",
     "@types/joi": "^17.2.3",
     "@types/node": "^20.7.1",
+    "jest": "^29.7.0",
     "typescript": "^5.2.2"
   },
   "dependencies": {
