--- platform-engineer-2/src/routes/index.ts
@@ -1,5 +1,0 @@
-import SaveImages from './save-images';
-
-export default {
-  saveImages: (body: any) => new SaveImages(body)
-}
