--- platform-engineer-2/src/routes/save-images/validation.ts
@@ -1,9 +1,0 @@
-import * as Joi from 'joi';
-
-export const schemaValidator = (body: any) => {
-  const schema = Joi.object({
-    jsonFilePath: Joi.string().required(),
-    outputFilePath: Joi.string().required(),
-  })
-  return schema.validate(body);
-}
